requirements.txt
praw==7.7.1
pandas>=2.1.0
pyarrow>=15.0.0
python-dotenv>=1.0.1
tenacity>=8.3.0
tqdm>=4.66.0
pyyaml>=6.0.1
